Terminada la práctica realizada con Jorge Delgado

+++ Segundo Semestre/Practica 1. Numero PI Iterativo/src/mates/Matematicas.java
@@ -6,12 +6,13 @@ public class Matematicas {
      * el número de puntos generados.
      */
     public static double generarNumeroPiIterativo(long pasos) {
-        int dentroDelCirculo = 0;
-        for (int i = 0; i < pasos; i++) {
+        long dentroDelCirculo = 0;
+        for (long i = 0; i < pasos; i++) {
             double x = Math.random(); // Coordenada X aleatoria entre 0 y 1
             double y = Math.random(); // Coordenada Y aleatoria entre 0 y 1
-
             // Si el punto cae dentro del círculo unitario, se aumenta el contador
+            // El uno se refiere a que el radio del círculo es 1
+            // El condicional es equivalente a la hipotenusa del triángulo formado por x e y
             if (x * x + y * y <= 1) {
                 dentroDelCirculo++;
             }
